Bug 797770 - Reconciliation report does not consider credit transactions
A few preparatory changes...

* calculated-cells augmented to allow 7 fields instead of 6. 7th field
defines whether cell is mergeable for subtotal-table. note only the
first element in calculated-cells is currently considered for the
subtotal-table accumulator.

* backward compatible upgrade to trep-engine's add-subtotal-row --
handles both 6-field and 7-field calculated-cells. 7th field, if
missing, is assumed to be #f.

* transaction report: unlock subtotal-table when amount='double in
addition to 'single.

And finally the bug fix itself:

* reconcile-report: fix subtotal-table accumulator to consider both Dr
and Cr