Place the cursor after the preedit.
Connect to the GtkEntry's preedit-changed signal and stash the preedit
length.

Uses the PangoLayout's text instead of the GtkEntry's because the former
includes the current preedit string and the latter does not. Add the
preedit length to the cursor position so that the cursor is drawn to the
right (left in RTL languages) of the preedit.