Rework command line option parsing to store values directly in class member variables
The program options libary has a convenience binding for boost::optional
to indicate whether an option is set or not.
Use this to store options passed on the command line directly
in variables for later use. This avoids the need to refer to options
in several locations using a fixed string (like 'help', 'help-gtk', 'nofile',...)

In addition drop a number of obsolete class member variables.
They were leftovers from the conversion to c++ and no longer used.