Fix g_signal_handler_disconnect: assertion in trace file
There is a difference in the way the sheet is closed between registers
and invoices. The gnc-date-cell-destroy is being called before the
gnc-item_edit_destroying for invoices and so when the item_edit tries
to do the signal disconnect on the popup_item it no longer exists.
Moving the gtk_widget_destroy before ledger_destroy fixes this.