Bug 797774 - Bill/Invoice due reminder confusion on CN?
The Bill/Invoice due reminder has a column header of 'CN?' which was
unclear of meaning which is 'Is this xxx a Credit Note'. A tooltip
was suggested but a better fix is to change the column to display the
'Type' so you would see Bill, Invoice or 'Credit Note' just like you
see in the find dialog.