Bug 797531 - Improve jumps to filtered registers
Currently if a jump to a filtered register is made, this could be from
a report, other register, reconcile window and transaction associations
and the destination split is not shown the jump will end up at the last
active cell. This could be confusing so add a test for the destination
split being visible and warn the user with an option to temporarily
clear the filter.