Bug 797791 - [Windows] list of command line options after 'gnucash.exe --help'?
This commit makes gnucash-cli a console application on Windows which
means it's output will be redirected to the connected console.
This works both in cmd.exe and in powershell.

gnucash itself remains a Windows GUI application and hence won't output
to console. To capture its output one needs to start it from cmd.exe
(not powershell!) as follows
'path-to-gnucash.exe' >'file-for-stdout' 2>'file-for-stderr'

'file-for-stdout' and 'file-for-stderr' are the names file two arbitrary
files the user has write access to. They don't have to exist beforehand.

Note 'file-for-stderr' can be set to '&1' (without the single quotes)
to redirect everything on stderr to wherever stdout goes.