Bug 797800 - [help screen items] clarification welcome
This fixes the presence of the 'input-file' command line option.
It's an implementation detail that wasn't meant to be listed in help.
The way to fix it is keeping two option_description variables. One
with all possible values to parse and one with only those to present
to the user