Bug 787295 - Allow UI jump from Business accounts to their relevant invoice/bill/voucher
Amendment - the current implementation will allow to click
on each split in a bussines account (payments and invoices)
but won't allow to jump from an invoice split in an income
or expense account. This amendment changes this such that
one can jump from any split whose transaction is an invoice
transaction. This allows jumping from income and expense accounts
if their splits are related to an invoice. It loses the ability
to jump from payment splits in a business account.
This is however more correct as payment transactions are not
necessarily associated with exactly one invoice (it can be
none, one or many).
If jumping from payment transactions should be supported as well
that will require more effort, namely if many invoices are
associated with the payment, the user should be asked which one
to open.