Bug 787295 - Allow UI jump from Business accounts to their relevant invoice/bill/voucher
Combine the best of both approaches and allow jumping to invoice if
either
- the current txn is an invoice transaction
- the current split has an invoice associated with it
Those are the only two conditions that allow us to uniquely
identify an invoice from a register.
A future extension could be to allow a jump if any split
of the current txn has an invoice associated with it.
This would need user interaction if more than one split
is linked to an invoice. A payment transaction for example
can pay more than one invoice and hence can have more than one
invoice related split.