To further increase the import speed, it's necessary to prevent any account commit to happen until the very end of the import (OK or Cancel), because account commits trigger very lengthy balance computations. For this, I call xaccAccountBeginEdit on all the accounts involved in the import, keeping a list of them so BeginEdit is called only once. At the end of the import, commit is called on all the accounts in the list. Note that when the user selects a target account for an imported transaction, xaccAccountBeginEdit is called on the target account, and it is added to the list.
Another area of improvement is avoiding re-checking all register transactions to verify whether
a given imported transaction has already been matched. Instead, a hash table of split online IDs
is computed once (per account), and verified for each incoming transactions.

Finally, the list of register transactions that are potential matches for the imported ones is
further pruned ahead of time to only keep transactions that do not have an online ID. This avoid
the repeated checks that were previously happening in the match-score loop.

With this, importing 6000 transactions into a 6000 split account becomes fairly fast (a few seconds
on my slowish machine).

There are still slow areas: If you select all 6000 imported transactions and assign a destination
account to all of them, the process is impossibly sluggish because of repeated path operations
(selections, freeing) in the tree view.
If you do not specify a target account for any of the 6000 imported transactions, an "imbalance"
account is used, but the xaccAccountBeginEdit mechanism isn't applied to it by the new code, so
each imported transaction will trigger a commit, and therefore a slow balance recomputation.

Remove use of xaccTransGetSplit