Multiple-account selection and delete.
This PR implements multi-selection of accounts multi-account (rubberbanding) delete. This is an ease-of-use improvement, especially for new users. Most of us don't typically change our accounts once our books are setup I imagine, but new users do.
Trying to keep this message brief.

Somebody in the past must have looked at multi-account actions because there already was functioning (but never called) code to get the list of selected accounts.

Only the Delete action is currently available when multiple accounts are selected in the account page. But I will follow up with another PR that implements the multi-account edit action.

- GC used to delete accounts on the spot if there was no transactions or subaccounts in them. I removed that: you always have to confirm.
- Once you've chosen what to do with the transactions and subbacounts, I check that the selected actions are compatible with all accounts selected for delete. I ask for confirmation, then I delete all accounts.
- Moving transactions used to be allowed only to an account of the same type as the deleted account. I'm not sure why this is strictly enforced, but I removed it as it does not seem warranted. I could add a warning to tell the user.