This PR implement multi-account editing.
The PR is meant as a continuation of multi_account_delete.
To make multi-account simpler I only allow multi-account editing of accounts under the same parent (for now).
The difficulty is to present the data to the user when the accounts have mismatched attributes. Generally, the UI indicates a mismatch (see below) and only when the user makes a selection is the mismatch removed and replaced by the user's selection. If the user does not change the mismatched attributes, they're not modified in the accounts.
Details:

- Mismatched strings are easy: show "mismatched".
- Mismatched flags are easy: the GTK button has a "inconsistent" state.
- Mismatched currency: I add a temporary fake currency "MULTIPLE" which is removed as soon as the user edits the currency.
- Mismatched colors: I don't have a good solution for that: the GTK color button does no offer a way to show "inconsistency". At the moment, I show gray, but that's not very good. Ideas welcome.
- Mismatched parents: Impossible given the current restriction (see above). Changing the parent is supported of course.
- Mismatched Account Types:
   . I switch to a multi-selection type view, and show the different types of the accounts as selected (so you'll see more than 1 type in blue)
   . As soon as the user selects a common type, I switch back to single-selection.

The entire thing is complicated by the fact that GTK does not differentiate (something_changed callbacks triggered when the code changes something internally), and (something_changed callback triggered because the user clicked something). Only when the user makes a selection should the mismatch be collapsed, not when callbacks are triggered internally. In addition, GTK does not have a way to temporary disable callbacks.
An example of this is if the user selects a different parent: this typically changes data in the account type tree/view but we do not want these to understood as if the user had made them.

The solution I'm using is a flag in the aw structure to indicate whether callbacks should be on or not.