Bug 797874 - new account UI error from register
If the path 'test:testy' is typed in to the transfer field and it does
not exist you are asked if you want to create it. The new account
dialog opens for 'test' which you create but this new 'test' account
does not show in the parent account tree for the second new account
dialog.

To fix this the existing account parent_tree is dropped and a new one
is use created which has the newly created parent account.