Bug 797971 - Inserting description text behaves oddly
This is best observed when trying to insert a character, say x when the
cursor is before the last character, the cursor will jump to the end
like... ABCD|E results in ABCDxE|

It can also affect when inserting/changing text that has come from the
quick fill the cursor may end up in the wrong place.

Changed a test condition to use the incoming new character length value
instead of the existing cell one.