Bug 345924 - Conditionally change Options dialog cancel button
When the options dialog is loaded, the main buttons are a sensitive
'Cancel' button and insensitive 'OK' and 'Apply' buttons. When changes
are made the 'OK' and 'Apply' buttons become sensitive and when pressed
the 'OK' button will make changes and close dialog whereas the 'Apply'
button will make changes and the dialog will remain open but with these
buttons insensitive leaving just 'Cancel' sensitive.

This could be confusing so this commit changes conditionally the text
label of the 'Cancel' button between 'Cancel' and 'Close'.