Bug 797944 - Crash when opening new file from existing one
If you have report pages that are left open when the current book is
'Saved' and then use the 'File->Open' to open a new book Gnucash may
crash. This was traced to the use of a g_idle_add to delay loading the
reports in the pages that was not being cleared when the first book was
closed.

To fix this the g_idle_add has been removed and the loading of the
report has been moved to the page focus function. By using a flag to
indicate when pages are being restored, the report loading is delayed
until the page is focused which has also decreased book loading time
especially if no report pages were the current page when the book was
closed.