Intercept the Ctrl+Alt+PgUp/Down for a report page
With the focus on the webkit webview, this key sequence can be used to
scroll the view. For the key sequence to work for tab page switching a
callback is connected to the 'key-press-event' for the webkit view and
checked for the sequence and passed to the notebook if required.