Change Preference dialog account separator validation
If the account separator is changed and is invalid the warning dialog
may be shown many times. To fix this the validation is done on closing
the preference dialog when a validation dialog will present two options,
The first is to reset the separator to the original value and close the
preference dialog and the second is to return to the 'Accounts' page of
the preference dialog.

The warning symbol is still displayed when the separator is changed
giving an indication of a problem.

This has stemmed from PR #766 and all this commit does is the minimum
recommended there.