Bug 644242 - Register Panning Problem
If the horizontal scrollbar is present for a register and you start to
tab from one cell to the next one that is not in view the sheet does
not scroll to keep the active cell in view.

This commit fixes that by getting the active cell horizontal offset and
width and comparing it to the sheet width and window width.