Combine multiple splits for the same transaction
Currently when you have multiple splits for the same account in the
same transaction they will appear in the register multiple times, one
for each split when viewed in basic view. This has been confusing for
users so this commit changes that so with the basic view, the splits
values for the account are added together.