Bug 797901 - File History not updated
To reproduce this, with an existing list of four recently opened files,
open a fifth file whose name must differ from those already in the list
and you will observe that the file history list is not updated.

This was down to when doing the 'File->Open', a call is made to
gnc_close_gui_component_by_session which closes the existing pages and
as part of that the preference callbacks for the history plugin page
were being removed along with some others like for tab placement.

To fix this added the removal inside an if statement that checks to see
if the window is quitting or not.