Bug 720564 - Search in General Ledger
If you search from a General Ledger the search results will update the
General Ledger and the only way to get back the contents is to close
and reopen the General Ledger.

This change makes it the same as any other register in that the search
results are opened in a new search page.