Add parent widget to gnc_gnome_help function
If there are any errors from starting help with gnc_gnome_help they are
displayed with gnc_error_dialog which defaults to setting the parent
window to that of GnuCash. This can be confusing if the dialog is
behind another window, say the preference window as it would appear
that Gnucash is unresponsive so add the GtkWindow parameter to
gnc_gnome_help and populate appropriately.