Bug 343647 - Keep scrolled report location
If you have a large report and have scrolled down to the bottom and
decide to tab to a different page and come back the report is back at
the top. This was due to reloading the report before focusing so change
this so it only loads once, if the report needs refreshing then there
is the reload toolbar button.