Extend new algorithm by a token selection algorithm
By this extension individual tokens can be selected/excluded from the
final probability calculation.

In this first implementation tokens are selected automatically, if there
is at least one conditional account probability P(A|T) for this token,
which is larger than a defined threshold.

The current threshold is quite high, i.e. only tokens which are (almost)
unique to one account are selected for final probability calculation.
This emulates the behaviour of the previous algorithm, whose results
also only depend on tokens, which are unique to one account.

Additionally the existing account probability threshold can be reduced
due to the new introduced threshold for individual tokens.