Increase the maximum threshold for date mismatch
Some french banks (that issue credit cards) have OFX files that count
the transaction as dated of the 5 of the month that follows the actual
expense. This means that the date offset between the invoice (that is at
the date of the expense) and the OFX file can be up to around 30-35
days.

So without this change, “usual” scores for such transactions are:
  +3 for cent-accurate amount
  -5 for date that mismatches by more than the max threshold
  +1 when lucky for matching the description or memo
thus leading to a score of -1, which can literally never match, as the
setting for reducing the minimal score for matches can only go down to
zero. And thus OFX matching would be impossible with the window, without
duplicating the transactions and then manually doing the matching
manually.

The only alternative to making this change that I could see would be to
parse the actual date, which the bank OFX includes in the description
field, and to compare with this date. It would probably be a more
precise solution, but would I think be pretty awfully bank-specific,
which in turn makes it a probably bad idea.