Bug 798149 - Account name change affects CSV saved import setting
The account full path was being used in the CSV saved settings so if
any account name is change in the path it would prevent recall of the
base account used. To fix this the account guid is saved instead and
when recalled the account is looked up first by guid, if this fails the
full path is checked which if successful immediately updates the saved
base account setting with the account guid for future use.