Bug 797621 - Sometimes freeze on Invoice Sales Tax Table
If there are no entries in the list, the return height of
popup_get_height was 0 but when a confirmation check is made on true
height it returned 1 and so try and queue a resize and hence a loop so
in popup_get_height return 1 if list is empty.

Also to prevent this from happening entirely, disable the toggle button
when there are no entries in the list. The entry will still be active
so new list entries can be created as before.