Add two new preferences for import
In the import-backend.c source file two constants were defined with a
comment about they should ideally be added to the user preference
dialog. This commit does that, the two preferences are
match-date-threshold which is the number of days a transaction match is
likely and match-date-not-threshold which is the number of days a
transaction match is considered not likely.