Bug 798159 - Keyboard shortcut bug to manage document link
The GtkFileChooserButton was grabbing focus on load but was not
responding to it, possibly due to it already been created but hidden so
change it to a normal button that starts a GtkFileChooserNative dialog.
Also stop grabbing focus on the use of the toggle buttons to aid
navigation by keyboard.