Bug 753283 - Highlight missing when returning from schedule edit.
When you return from editing a scheduled transaction the highlight is
missing in the tree view. This is partly due to the model being
recreated when ever a change is made to a scheduled transaction. To
overcome this, the selected schedules are added to a GList which is
used in a new call back signal that has been added to indicate when the
model is populated. When this happens, the model is scanned for the
schedules that are in the GList so they can be reselected.