Bug 798195 - No option in the user interface to set the opening balance of an existing account
Previously, opening balances could only be entered through the hierarchy
assistant or when creating a new account. This excludes the entry of opening
balances for accounts that were already included in the chart of accounts
when the file was created, but were not used. In these cases, it was
necessary to manually create an entry in the corresponding account.

By enabling the opening balance page when editing accounts when there are
no transactions, the same procedure as for new creation is enabled.