Provide a selection option for the opening balance account in the hierarchy assistant
With the hierarchy assistant, a chart of accounts can be compiled from a
set of individual templates. Since some templates contain an opening
balance account, it may happen that several opening balance accounts
are available, from which the user has to select one.

To do this, this commit adds a new column to the final account tree to
select a single opening balance account.

When the wizard exits, it checks if multiple opening balance accounts
are selected. In this case, the wizard cannot be completed and the user
has the option to correct the case or cancel the wizard.

see #820
see https://bugs.gnucash.org/show_bug.cgi?id=797836#c12