Incorrect validation for GNCAmountEdit
gnc_amount_edit_expr_is_valid can return 0 for validation OK, 1 for an
error with no position information and also the position of that error.
If the error was at position 0, that will look like the validation was
OK so add 1000 to the position information so there is no confusion.