Change GNCAmountEdit to be based on a GtkBox
To indicate that the expression entered has failed change the widget it
is based on to a box and add a warning triangle image with tooltip that
is made visible when the validation has failed. This is less intrusive
but if required the error message can be retrieved and used for a
warning dialog if that is in keeping with other dialog checks.