Incorrect validation for GNCAmountEdit
gnc_amount_edit_expr_is_valid can return 0 for validation OK, 1 for an
error with no position information and also the position of that error.
If the error was at position 0, that will look like the validation was
OK so instead return 1 and use a GError parameter using the position as
the code and setting the message to the common format used. Also as
gnc_amount_edit_evaluate could also benefit from using the GError add
that as a parameter and initially set all occurrences of use to NULL.