Bug 798202 - Register editor Black on Black background
The GtkEntry of the register editor was not having the appropriate
style class gnc-class-register-cursor or gnc-class-user-register-cursor
applied resulting in black text on a black background when the theme
"Breeze Dark" was used.

To fix this, in function gnucash_sheet_refresh_from_prefs add setting
the appropriate style classes for the entry after first removing the
existing "gnc-class-" classes.