Type ahead not working properly on subsequent key strokes
When in the transfer cell and you are searching for the 'fuel' account,
entering the 'f' would bring up all accounts containing an 'f' but when
the next letter 'u' is entered it would show all accounts containing a
'u'. This is down to the transfer cell being updated to early from a
gnc_item_list_select which is being used for the quick fill so test for
quick fill use first.