Bug 798212 - right-click the register down-arrow
If the toggle buttons on the register sheet, used for date and transfer
cells, is clicked with a right mouse the focus would change to the
first transaction.

To fix add a 'button-press-event' callback to intercept the right mouse
and also prevent the main register sheet menu from showing when an
item_edit pop up is showing.