Bug 798234 - Cut Transaction discards the reference to the description/memo strings so that Paste Transaction will paste uninitialised data (or other strings)
When a transaction is cut, any globally unique description and memo strings
that are referenced by FloatingTxn/FloatingSplit will refer to memory that
has been freed. This results in uninitialised data (or other strings) being
inserted when the transaction is pasted.

These strings are in the string cache so take a new reference to them when
copying them to the FloatingTxn/FloatingSplit.

Fix memory leaks of FloatingTxn and FloatingSplit when they're used
temporarily.

Fix memory leak in FloatingTxn by freeing the m_splits list too.