Bug 798205 Do not exclude from import a transaction that has an FITID
which matches an already existing split. This is because it
(1) may be a coincidence and therefore the transaction needs to be added
or
(2) could be on a transfer from the account being imported to another
bank account which has already been imported, and so needs to be
matched, not added.