[dialog-lot-viewer] g_free a GList* properly
g_list_free requires the argument is the head of a GList. Calling
g_list_reverse makes the filtered_list points to the last
element. Assigning filtered_list to the result of g_list_reverse
ensures it still points to the head, allowing g_list_free to free the
list properly.